app/requirement.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
openai==1.57.0
pinecone-client==5.0.1
requests==2.32.3
pydantic==2.7.4
python-dotenv==1.0.1
